requirements.txt
Django
libsass
django-compressor
psycopg2
django-autoslug
singlemodeladmin
Unidecode
django-ipware
Pillow
django-admin-sortable2
